# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "name": "API Support",
      "url": "https://github.com/fransfilastap/kontak"
    },
    "title": "Kontak API",
    "version": "1.0",
    "description": "WhatsApp-to-REST-API bridge",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
...
    "/admin/clients": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get all devices for the authenticated user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "devices"
        ],
        "summary": "List devices",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Register a new WhatsApp device",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "devices"
        ],
        "summary": "Register device",
        "parameters": [
          {
            "description": "Device data",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/clients/{client_id}/connect": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Connect to a WhatsApp device",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "devices"
        ],
        "summary": "Connect device",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/pkg_http.DeviceConnectionResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.DeviceConnectionResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.DeviceConnectionResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.DeviceConnectionResponse"
            }
          }
        }
      }
    },
    "/admin/clients/{client_id}/disconnect": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Disconnect from a WhatsApp device",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "devices"
        ],
        "summary": "Disconnect device",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/pkg_http.DeviceConnectionResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/clients/{client_id}/qr": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get QR code for device pairing",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "devices"
        ],
        "summary": "Get QR code",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/clients/{client_id}/status": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get the connection status of a device",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "devices"
        ],
        "summary": "Get connection status",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/v1/chats": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Send a text message via WhatsApp",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "messages"
        ],
        "summary": "Send message",
        "parameters": [
          {
            "description": "Message data",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pkg_http.SendMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/pkg_http.SendMessageResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/v1/chats/media": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Send a media message via WhatsApp",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "messages"
        ],
        "summary": "Send media message",
        "parameters": [
          {
            "type": "string",
            "description": "Client ID",
            "name": "client_id",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "Mobile number",
            "name": "mobile_number",
            "in": "formData",
            "required": true
          },
          {
            "type": "file",
            "description": "Media file",
            "name": "media_url",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "Caption",
            "name": "caption",
            "in": "formData"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/pkg_http.GenericResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/v1/chats/template": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Send a message using a template",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "messages"
        ],
        "summary": "Send template message",
        "parameters": [
          {
            "description": "Template message data",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pkg_http.SendTemplateMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/pkg_http.GenericResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "pkg_http.DeviceConnectionResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "server_error": {
          "type": "boolean"
        }
      }
    },
    "pkg_http.ErrorResponse": {
      "type": "object",
      "properties": {
        "context": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "pkg_http.GenericResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "pkg_http.MessageType": {
      "type": "string",
      "enum": [
        "text",
        "image"
      ],
      "x-enum-varnames": [
        "MessageTypeText",
        "MessageTypeMediaImage"
      ]
    },
...
    "pkg_http.SendMessageRequest": {
      "type": "object",
      "required": [
        "client_id",
        "mobile_number",
        "text",
        "type"
      ],
      "properties": {
        "client_id": {
          "type": "string"
        },
        "mobile_number": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/pkg_http.MessageType"
        }
      }
    },
    "pkg_http.SendMessageResponse": {
      "type": "object",
      "properties": {
        "message_id": {
          "type": "string"
        }
      }
    },
...
    "pkg_http.SendTemplateMessageRequest": {
      "type": "object",
      "required": [
        "deviceId",
        "templateId",
        "to"
      ],
      "properties": {
        "deviceId": {
          "type": "string"
        },
        "templateId": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "variables": {
          "type": "object",
          "additionalProperties": true
        }
      }
    }
  },
  "host": "localhost:8080",
  "basePath": "/"
}